packages/scrapers/requirements.txt
scrapy>=2.11
scrapy-playwright>=0.0.33
pydantic>=2.0
sqlalchemy>=2.0
psycopg2-binary
python-dotenv
fake-useragent
